requirements.txt
presidio-analyzer
presidio-anonymizer
PyPDF2
pytesseract
Pillow

